move scanner quiz_util to util/
