+++ depobs/clients/github.py
@@ -34,7 +34,7 @@ from depobs.scanner.models.github import (
     get_next_requests,
     MISSING,
 )
-from depobs.scanner.quiz_util import raw_result_to_dict
+from depobs.util.quiz_util import raw_result_to_dict
 from depobs.util.traceback_util import exc_to_str
 
 log = logging.getLogger(__name__)
